+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [1.8.1]  / 2023-05-13
 ### Updated
 - Update `GetFileVersionInfo` to find `exe` file.
+- Update `SignProject` to sign `Name`, `AssemblyName`, or `*`.
 
 ## [1.8.0]  / 2023-03-11
 ### Updated